+++ src/main/java/duke/command/DeadlineCommand.java
@@ -1,10 +1,10 @@
-package duke.commands;
+package duke.command;
 
 import duke.exception.DukeException;
 import duke.storage.Storage;
-import duke.tasks.Deadline;
-import duke.tasks.Task;
-import duke.tasks.TaskList;
+import duke.task.Deadline;
+import duke.task.Task;
+import duke.task.TaskList;
 import duke.ui.Ui;
 
 import java.time.LocalDateTime;
