+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,23 +1,22 @@
-package duke.commands;
+package duke.command;
 
 import duke.exception.DukeException;
 import duke.storage.Storage;
-import duke.tasks.Task;
-import duke.tasks.TaskList;
+import duke.task.Task;
+import duke.task.TaskList;
 import duke.ui.Ui;
 
-public class DoneCommand extends Command {
+public class DeleteCommand extends Command {
     private int index;
 
-    public DoneCommand(int index) {
+    public DeleteCommand(int index) {
         this.index = index;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
-        Task tempTask = tasks.getByIndex(index);
-        tempTask.markDone();
+        Task tempTask = tasks.deleteTaskByIndex(index);
         storage.saveAll(tasks);
-        ui.showDone(tempTask);
+        ui.showDeleteTask(tempTask, tasks.getSize());
     }
 }
