+++ src/main/java/duke/command/ExitCommand.java
@@ -1,7 +1,7 @@
-package duke.commands;
+package duke.command;
 
 import duke.storage.Storage;
-import duke.tasks.TaskList;
+import duke.task.TaskList;
 import duke.ui.Ui;
 
 public class ExitCommand extends Command {
