+++ src/main/java/duke/command/ListCommand.java
@@ -1,8 +1,8 @@
-package duke.commands;
+package duke.command;
 
 import duke.exception.DukeException;
 import duke.storage.Storage;
-import duke.tasks.TaskList;
+import duke.task.TaskList;
 import duke.ui.Ui;
 
 public class ListCommand extends Command {
