+++ src/main/java/duke/parser/Parser.java
@@ -1,6 +1,6 @@
 package duke.parser;
 
-import duke.commands.*;
+import duke.command.*;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
