+++ src/main/java/duke/task/Event.java
@@ -1,4 +1,4 @@
-package duke.tasks;
+package duke.task;
 
 import java.time.LocalDateTime;
