+++ src/main/java/duke/task/TaskType.java
@@ -1,4 +1,4 @@
-package duke.tasks;
+package duke.task;
 
 public enum TaskType {
     T, D, E, X
