+++ tutorials/javaFxTutorialPart3.md
@@ -228,7 +228,7 @@ You have successfully implemented a fully functional GUI for duke.Duke!
 
 ## Exercises
 
-1. While the GUI looks similar to the mockup, there are still parts that need to be refined. Try your hand at some of these duke.tasks:
+1. While the GUI looks similar to the mockup, there are still parts that need to be refined. Try your hand at some of these duke.task:
    * Add padding between each DialogBox
    * Add padding between each ImageView and its Label
    * Clip the ImageView into a circle
